% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R
\name{didml-deprecated}
\alias{didml-deprecated}
\alias{fddml}
\alias{fddml_nuisance}
\alias{fddml_wald}
\alias{fddml_tc}
\alias{fddml_trim}
\alias{fddml_inference}
\title{Deprecated fddml Functions}
\usage{
fddml(Y, D, G, Ti, X, estimand = "both", ...)

fddml_nuisance(...)

fddml_wald(...)

fddml_tc(...)

fddml_trim(...)

fddml_inference(...)
}
\arguments{
\item{Y}{Numeric vector of outcomes.}

\item{D}{Numeric vector of treatment.}

\item{G}{Binary vector of group assignment (0/1).}

\item{Ti}{Binary vector of time period (0/1).}

\item{X}{Numeric matrix of covariates.}

\item{estimand}{Character: passed as \code{estimator} to \code{didml()}.}

\item{...}{Additional arguments passed to the replacement function.}
}
\description{
These functions are deprecated in favor of the \code{didml} family.
They will be removed in a future release.
}
