% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.didml}
\alias{print.didml}
\title{Print a didml object}
\usage{
\method{print}{didml}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{didml} object.}

\item{digits}{Number of significant digits (default 4).}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a didml object
}
