% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.didml}
\alias{print.summary.didml}
\title{Print a summary.didml object}
\usage{
\method{print}{summary.didml}(x, ...)
}
\arguments{
\item{x}{A \code{summary.didml} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a summary.didml object
}
