% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.didml}
\alias{summary.didml}
\title{Summary of a didml object}
\usage{
\method{summary}{didml}(object, ...)
}
\arguments{
\item{object}{A \code{didml} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
An object of class \code{summary.didml}, printed via
\code{\link{print.summary.didml}}.
}
\description{
Summary of a didml object
}
